requirements.txt
arrow==1.3.0
asgiref==3.7.2
attrs==23.1.0
autobahn==23.6.2
Automat==22.10.0
blessed==1.20.0
bx_django_utils==70
bx_py_utils==90
certifi==2024.8.30
cffi==1.16.0
constantly==23.10.4
cryptography==41.0.7
da-vinci==0.4.0
daphne==4.0.0
distlib==0.3.9
Django==5.0
django-admin-sortable2==2.1.10
django-auto-prefetching==0.2.12
django-ckeditor==6.7.0
django-cleanup==8.0.0
django-dynamic-preferences==1.16.0
django-filter==23.5
django-groups-manager==1.3.0
django-huey-monitor==0.9.0
django-js-asset==2.2.0
django-mptt==0.15.0
django-multiselectfield==0.1.12
django-picklefield==3.1
django-q==1.3.9
django-simple-history==3.4.0
django-thumbnails==0.7.0
djangorestframework==3.14.0
drf-spectacular==0.27.0
drf-writable-nested==0.7.0
filelock==3.16.1
huey==2.5.0
hyperlink==21.0.0
idna==3.6
incremental==22.10.0
inflection==0.5.1
jsonschema==4.20.0
jsonschema-specifications==2023.11.2
lxml==4.9.4
packaging==24.2
persisting-theory==1.0
Pillow==10.1.0
pipenv==2024.4.0
platformdirs==4.3.6
psycopg2-binary==2.9.9
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycparser==2.21
pyOpenSSL==23.3.0
python-dateutil==2.8.2
python-docx==1.1.0
python-dotenv==1.0.0
python-stdnum==1.19
pytz==2023.3.post1
PyYAML==6.0.1
redis==3.5.3
referencing==0.32.0
rpds-py==0.15.2
service-identity==23.1.0
shortuuid==1.0.11
six==1.16.0
sqlparse==0.4.4
Twisted==23.10.0
txaio==23.1.1
types-python-dateutil==2.8.19.14
typing_extensions==4.9.0
uritemplate==4.1.1
virtualenv==20.27.1
wcwidth==0.2.12
zope.interface==6.1
